-Fix for Windows warnings -Correct size_type assigned

+++ src/compiler/parser/jsoniq_scanner.cpp
@@ -1,6 +1,7 @@
-#line 2 "jsoniq_scanner.yy.cpp"
+#include "stdafx.h"
+//#line 2 "jsoniq_scanner.yy.cpp"
 
-#line 4 "jsoniq_scanner.yy.cpp"
+//#line 4 "jsoniq_scanner.yy.cpp"
 
 #define  YY_INT_ALIGNED short int
 
@@ -6242,7 +6243,7 @@ int yyFlexLexer::LexerInput( char* buf, int max_size )
 	if ( yyin->bad() )
 		return -1;
 	else
-		return yyin->gcount();
+		return static_cast<int>(yyin->gcount());
 #endif
 }
